requirements.txt
dnspython>=2.4.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0
pytest>=7.0.0
